MedicalEval/Question-answering_Score/med-flamingo/requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate@4d583ad6a1f13d1d7617e6a37f791ec01a68413a
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
# backports.zoneinfo==0.2.1
bert-score==0.3.13
bitsandbytes==0.39.0
blinker==1.6.2
braceexpand==0.1.7
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.9.0
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.9.1
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.0
einops-exts==0.0.4
entrypoints==0.4
evaluate==0.4.0
executing==1.2.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
h5py==3.8.0
hjson==3.1.0
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lit==16.0.5.post0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open-flamingo==0.0.2
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
peft==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.0
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.28.2
responses==0.18.0
rich==13.4.2
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
sentry-sdk==1.18.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.24.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
timm==0.9.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers@01734dba842c29408c96caa5c345c9e415c7569b
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
wandb==0.14.0
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.48
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
